implemented adding new contact from form
