+++ src/app.js
@@ -5,25 +5,6 @@ import ApplicationView from 'app/views/application_view';
 import Rolodex from 'app/collections/rolodex';
 import RolodexView from 'app/views/rolodex_view';
 
-var contactList = [
-  {
-    name: "Jenny",
-    email: "jenny@tutone.com",
-    phone: "867-5309"
-  }, {
-    name: "Jessica Rabbit",
-    email: "jrabbit@toontown.com",
-    phone: "555-0123"
-  }, {
-    name: "Elmer Fudd",
-    email: "killdawabbit@hunter.com",
-    phone: "KLondike 6-5000"
-  }, {
-    name: "Lisa Simpson",
-    email: "saxophone.nerd@sprinfield.com",
-    phone: "555-0420"
-  }
-];
 
 $(document).ready(function() {
   var application = new Application();
@@ -33,11 +14,6 @@ $(document).ready(function() {
     model: application
   });
 
-  var myRolodex = new Rolodex(contactList);
-  var rolodexDisplay = new RolodexView(
-    {
-      el: $("#contact-cards"),
-      model: myRolodex
-    });
-  rolodexDisplay.render();
+
+  appView.render();
 });
